Keep monitoring and telemetry orchestrators small
Split reporting renderers and telemetry query aggregation into composition helpers while preserving existing public import paths and behavior.

Constraint: Issue #613 requires quality-reporter.ts and telemetry-repository.ts to stay under 500 lines with quality/telemetry specs green.

Rejected: Leave query and renderer logic inline | fails the orchestrator size target and keeps unrelated responsibilities coupled.

Confidence: high

Scope-risk: moderate

Directive: Preserve telemetry-repository.ts and quality-reporter.ts as public facades; place new query/render behavior in adjacent helpers.

Tested: npm run type-check; npm run lint; targeted quality/telemetry/admin route specs; npm test; graphify rebuild; git diff --check.

Not-tested: production deployment smoke.
